fix(mobile): clean regex parsers
- Reduce parser complexity

- Split email persistence helper
